Refactoring metric trackers
