+++ test/test_segmentationtracker.py
@@ -7,7 +7,7 @@ import sys
 ROOT = os.path.join(os.path.dirname(os.path.realpath(__file__)), "..")
 sys.path.append(ROOT)
 
-from metrics.metrics_tracker import SegmentationTracker
+from metrics.segmentation_tracker import SegmentationTracker
 
 
 class MockDataset:
